Refactor: move util functions to Router class
