+++ src/index.js
@@ -1,5 +1,4 @@
 import { Router } from "./router";
-import { getPathParams } from "./utils/handleParams";
 import { Base } from "./components/Base";
 import { ListPage } from "./pages/List";
 import { DetailPage } from "./pages/Detail";
@@ -14,7 +13,7 @@ class App extends Base {
       { path: "/post", page: new ListPage() },
       { path: "/post/:id", page: new DetailPage() },
     ];
-    this.renderPostListPage(); //temporal test code
+    // this.renderPostListPage(); //temporal test code
     this.init();
   }
